requirements.txt
flask
authlib
flask-bcrypt
pymongo
python-dotenv
pyjwt
requests